allow re-use of OpenOptions in other crates
The `OpenOptions` struct's functionality of mapping the builder pattern APIs to
libc-level flags flags is useful outside of `tokio-uring`.

Our particular use case is an alternative approach to tokio + io_uring
which will be open-sourced soon. We are already re-using the `IoBuf`
trait there, so, it makes sense to also re-use `OpenOptions`.

So, this PR adds an extension trait for `OpenOptions` that allows producing
an `io_uring::squeue::Entry`.

(This is a back-port of https://github.com/neondatabase/tokio-uring/pull/1
to tokio-uring v0.4.0, the latest official release).